Publishing!!
Getting everything ready for publishing. Right now it will publish to IT\MarketingEventsApplication\.

Updates will need to go to IT\MarketingEventsApplication\Updates\.

Need to get the files needed to run in the correct place before it will be able to install.

When this is figured out, will do the same to the dev channel.